App GUI/requirements.txt
customtkinter~=4.3.0
pyserial~=3.5